+++ derive_builder/src/options/mod.rs
@@ -39,9 +39,7 @@ pub fn struct_options_from(ast: &syn::MacroInput) -> (StructOptions, OptionsBuil
 }
 
 ///  Get the `FieldOptions` for a field with respect to some custom default values.
-pub fn field_options_from(f: syn::Field,
-                          defaults: &OptionsBuilder<FieldMode>)
-                          -> FieldOptions {
+pub fn field_options_from(f: syn::Field, defaults: &OptionsBuilder<FieldMode>) -> FieldOptions {
     OptionsBuilder::<FieldMode>::parse(f).with_defaults(defaults).into()
 }
 
@@ -56,6 +54,7 @@ pub struct OptionsBuilder<Mode> {
     /// Takes precedence over `setter_prefix`
     setter_name: Option<String>,
     setter_vis: Option<syn::Visibility>,
+    field_vis: Option<syn::Visibility>,
     default_expression: Option<DefaultExpression>,
     setter_into: Option<bool>,
     try_setter: Option<bool>,
@@ -83,6 +82,7 @@ impl<Mode> From<Mode> for OptionsBuilder<Mode> {
             setter_name: None,
             setter_vis: None,
             try_setter: None,
+            field_vis: None,
             default_expression: None,
             setter_into: None,
             no_std: None,
@@ -91,8 +91,8 @@ impl<Mode> From<Mode> for OptionsBuilder<Mode> {
     }
 }
 
-impl<Mode> OptionsBuilder<Mode> where
-    Mode: OptionsBuilderMode
+impl<Mode> OptionsBuilder<Mode>
+    where Mode: OptionsBuilderMode
 {
     impl_setter!{
         ident: setter_enabled,
@@ -104,6 +104,12 @@ impl<Mode> OptionsBuilder<Mode> where
         ident: builder_pattern,
         desc: "builder pattern",
         map: |x: BuilderPattern| { x },
+    }
+
+    impl_setter!{
+        ident: field_public for field_vis,
+        desc: "field visibility",
+        map: |x: bool| { if x { syn::Visibility::Public } else { syn::Visibility::Inherited } },
     }
 
     impl_setter!{
@@ -148,8 +154,8 @@ impl<Mode> OptionsBuilder<Mode> where
         map: |x: String| { x },
     }
 
-    pub fn parse_attributes<'a, T>(&mut self, attributes: T) -> &mut Self where
-        T: IntoIterator<Item=&'a syn::Attribute>
+    pub fn parse_attributes<'a, T>(&mut self, attributes: T) -> &mut Self
+        where T: IntoIterator<Item = &'a syn::Attribute>
     {
         trace!("Parsing attributes.");
         for attr in attributes {
@@ -230,7 +236,7 @@ impl<Mode> OptionsBuilder<Mode> where
                 self.setter_public(false)
             },
             "setter" => {
-                // setter implicitly enabled
+        // setter implicitly enabled
                 self.setter_enabled(true)
             },
             "try_setter" => {
@@ -295,16 +301,12 @@ impl<Mode> OptionsBuilder<Mode> where
 
     /// e.g `setter(skip)` in `#[builder(setter(skip))]`
     #[allow(non_snake_case)]
-    fn parse_builder_options_list(
-        &mut self,
-        ident: &syn::Ident,
-        nested: &[syn::NestedMetaItem]
-    ) {
+    fn parse_builder_options_list(&mut self, ident: &syn::Ident, nested: &[syn::NestedMetaItem]) {
         trace!("Parsing list `{}({:?})`", ident.as_ref(), nested);
         match ident.as_ref() {
             "setter" => {
                 self.parse_setter_options(nested);
-                // setter implicitly enabled
+        // setter implicitly enabled
                 if self.setter_enabled.is_none() {
                     self.setter_enabled(true);
                 }
@@ -312,18 +314,34 @@ impl<Mode> OptionsBuilder<Mode> where
             "build_fn" => {
                 self.mode.parse_build_fn_options(nested)
             }
+            "field" => {
+                self.parse_field_options(nested);
+            }
             _ => {
                 panic!("Unknown option `{}` {}.", ident.as_ref(), self.where_diagnostics())
             }
         }
     }
 
+    fn parse_field_options(&mut self, nested: &[syn::NestedMetaItem]) {
+        trace!("Parsing field options.");
+        for x in nested {
+            match *x {
+                syn::NestedMetaItem::MetaItem(syn::MetaItem::Word(ref ident)) => {
+                    match ident.as_ref() {
+                        "private" => self.field_public(false),
+                        "public" => self.field_public(true),
+                        _ => panic!("Unknown field word `{:?}`. {}", ident, self.where_diagnostics())
+                    }
+                },
+                _ => panic!("Unknown field option `{:?}`. {}", x, self.where_diagnostics())
+            }
+        }
+    }
+
     /// e.g `skip` in `#[builder(setter(skip))]`
     #[allow(non_snake_case)]
-    fn parse_setter_options(
-        &mut self,
-        nested: &[syn::NestedMetaItem]
-    ) {
+    fn parse_setter_options(&mut self, nested: &[syn::NestedMetaItem]) {
         trace!("Parsing setter options.");
         for x in nested {
             match *x {
@@ -400,11 +418,7 @@ impl<Mode> OptionsBuilder<Mode> where
 
     /// e.g `setter(skip)` in `#[builder(setter(skip))]`
     #[allow(non_snake_case)]
-    fn parse_setter_options_list(
-        &mut self,
-        ident: &syn::Ident,
-        nested: &[syn::NestedMetaItem]
-    ) {
+    fn parse_setter_options_list(&mut self, ident: &syn::Ident, nested: &[syn::NestedMetaItem]) {
         trace!("Setter Options - Parsing list `{}({:?})`", ident.as_ref(), nested);
         match ident.as_ref() {
             _ => {